+++ src/bundle/Resources/config/services/dashboard.yml
@@ -40,6 +40,7 @@ services:
             - { name: ezplatform.tab, group: dashboard-everyone }
 
     EzSystems\EzPlatformAdminUi\Tab\Dashboard\PagerContentToDataMapper:
+        parent: EzSystems\EzPlatformAdminUi\Search\AbstractPagerContentToDataMapper
         autowire: true
         public: false
