+++ src/bundle/Resources/config/services/search.yml
@@ -4,11 +4,6 @@ services:
     autowire: true
     public: false
 
-  EzSystems\EzPlatformAdminUi\Search\PagerContentToDataMapper:
-    parent: EzSystems\EzPlatformAdminUi\Search\AbstractPagerContentToDataMapper
-    autowire: true
-    public: false
-
   EzSystems\EzPlatformAdminUi\Search\PagerSearchContentToDataMapper:
     parent: EzSystems\EzPlatformAdminUi\Search\AbstractPagerContentToDataMapper
     autowire: true
