+++ src/lib/Tab/Dashboard/MyMediaTab.php
@@ -11,7 +11,6 @@ namespace EzSystems\EzPlatformAdminUi\Tab\Dashboard;
 use eZ\Publish\API\Repository\SearchService;
 use eZ\Publish\Core\Pagination\Pagerfanta\ContentSearchAdapter;
 use eZ\Publish\Core\QueryType\QueryType;
-use EzSystems\EzPlatformAdminUi\Search\PagerContentToDataMapper;
 use EzSystems\EzPlatformAdminUi\Tab\AbstractTab;
 use EzSystems\EzPlatformAdminUi\Tab\OrderedTabInterface;
 use Pagerfanta\Pagerfanta;
@@ -20,7 +19,7 @@ use Twig\Environment;
 
 class MyMediaTab extends AbstractTab implements OrderedTabInterface
 {
-    /** @var \EzSystems\EzPlatformAdminUi\Search\PagerContentToDataMapper */
+    /** @var \EzSystems\EzPlatformAdminUi\Tab\Dashboard\PagerContentToDataMapper */
     protected $pagerContentToDataMapper;
 
     /** @var \eZ\Publish\API\Repository\SearchService */
@@ -32,7 +31,7 @@ class MyMediaTab extends AbstractTab implements OrderedTabInterface
     /**
      * @param \Twig\Environment $twig
      * @param \Symfony\Component\Translation\TranslatorInterface $translator
-     * @param \EzSystems\EzPlatformAdminUi\Search\PagerContentToDataMapper $pagerContentToDataMapper
+     * @param \EzSystems\EzPlatformAdminUi\Tab\Dashboard\PagerContentToDataMapper $pagerContentToDataMapper
      * @param \eZ\Publish\API\Repository\SearchService $searchService
      * @param \eZ\Publish\Core\QueryType\QueryType $mediaSubtreeQueryType
      */
