+++ src/lib/Tab/Dashboard/PagerContentToDataMapper.php
@@ -8,53 +8,44 @@ declare(strict_types=1);
 
 namespace EzSystems\EzPlatformAdminUi\Tab\Dashboard;
 
-use eZ\Publish\API\Repository\ContentService;
-use eZ\Publish\API\Repository\ContentTypeService;
-use eZ\Publish\API\Repository\LanguageService;
-use eZ\Publish\API\Repository\UserService;
-use eZ\Publish\Core\Helper\TranslationHelper;
-use eZ\Publish\Core\MVC\Symfony\Locale\UserLanguagePreferenceProviderInterface;
+use EzSystems\EzPlatformAdminUi\Search\AbstractPagerContentToDataMapper;
+use Pagerfanta\Pagerfanta;
 
-/**
- * @deprecated This class has been moved to \EzSystems\EzPlatformAdminUi\Search\PagerContentToDataMapper
- */
-class PagerContentToDataMapper extends \EzSystems\EzPlatformAdminUi\Search\PagerContentToDataMapper
+class PagerContentToDataMapper extends AbstractPagerContentToDataMapper
 {
-    /** @var \eZ\Publish\API\Repository\ContentService */
-    protected $contentService;
-
-    /** @var \eZ\Publish\API\Repository\ContentTypeService */
-    protected $contentTypeService;
-
-    /** @var \eZ\Publish\API\Repository\UserService */
-    protected $userService;
-
     /**
-     * @param \eZ\Publish\API\Repository\ContentService $contentService
-     * @param \eZ\Publish\API\Repository\ContentTypeService $contentTypeService
-     * @param \eZ\Publish\API\Repository\UserService $userService
-     * @param \eZ\Publish\Core\MVC\Symfony\Locale\UserLanguagePreferenceProviderInterface $userLanguagePreferenceProvider
-     * @param \eZ\Publish\Core\Helper\TranslationHelper $translationHelper
-     * @param \eZ\Publish\API\Repository\LanguageService $languageService
+     * @param \Pagerfanta\Pagerfanta $pager
+     *
+     * @return array
      */
-    public function __construct(
-        ContentService $contentService,
-        TranslationHelper $translationHelper,
-        ContentTypeService $contentTypeService,
-        UserService $userService,
-        UserLanguagePreferenceProviderInterface $userLanguagePreferenceProvider,
-        LanguageService $languageService
-    ) {
-        $this->contentService = $contentService;
-        $this->contentTypeService = $contentTypeService;
-        $this->userService = $userService;
+    public function map(Pagerfanta $pager): array
+    {
+        $data = [];
+        $contentTypeIds = [];
 
-        parent::__construct(
-            $translationHelper,
-            $contentTypeService,
-            $userService,
-            $userLanguagePreferenceProvider,
-            $languageService
-        );
+        foreach ($pager as $content) {
+            /** @var \eZ\Publish\API\Repository\Values\Content\Content $content */
+            $contentInfo = $content->contentInfo;
+
+            $contentTypeIds[] = $contentInfo->contentTypeId;
+            $data[] = [
+                'content' => $content,
+                'contentTypeId' => $contentInfo->contentTypeId,
+                'contentId' => $content->id,
+                'name' => $this->translationHelper->getTranslatedContentName($content),
+                'language' => $contentInfo->mainLanguageCode,
+                'contributor' => $this->getContributor($contentInfo),
+                'version' => $content->versionInfo->versionNo,
+                'content_type' => $content->getContentType(),
+                'modified' => $content->versionInfo->modificationDate,
+                'initialLanguageCode' => $content->versionInfo->initialLanguageCode,
+                'content_is_user' => $this->isContentIsUser($content),
+                'available_enabled_translations' => $this->getAvailableTranslations($content, true),
+            ];
+        }
+
+        $this->setTranslatedContentTypesNames($data, $contentTypeIds);
+
+        return $data;
     }
 }
